Optimization Algorithm to Design Bioretentions
In this zipped file, two functions are presented. First, the fitness function of the optimization algorithm is presented. The second file defines the parameters for the optimization problem. The ultimate goal is to define a near-optimal design that satisfy local constraints as minimum peak flow reduction, maximum ponding depth, and other defined by the parameters in the optimization problem.